Correção da Tela de alteração
Correção da Tela de alteração do usuário

+++ Mygle.Grafico/TelaUsuarioResumo.cs
@@ -51,10 +51,10 @@ namespace Mygle.Grafico
 
         private void TelaUsuarioResumo_Shown(object sender, EventArgs e)
         {
-            Decimal Total = 0;
+            Double Total = 0;
             foreach (Venda venda in Program.Gerenciador.TodasAsVendas())
             {
-                Total+=venda.ValorTotal;
+                Total += venda.ValorTotal;
             }
             this.lbValorVendido.Text = Total.ToString();
         }
